fix Long to int

+++ src/main/java/com/swp391/dto/request/ReminderRequest.java
@@ -13,6 +13,6 @@ import java.time.LocalDate;
 public class ReminderRequest {
     LocalDate recoverTime;
     Boolean isSent;
-    Long memberId;
-    Long adminId;
+    int memberId;
+    int adminId;
 }
